+++ include/RAJA/pattern/kernel/Lambda.hpp
@@ -95,14 +95,7 @@ struct StatementExecutor<statement::Lambda<LoopIndex, Args...>> {
     //Convert SegList, ParamList into Seg, Param types, and store in a list
     using targList = typename parser<camp::list<Args...>>::type;
 
-    //Create a tuple with the appropriate lambda arguments
-    auto argTuple = call_extractor<targList>::make_tuple(data);
-
-    //Invoke the lambda with custom arguments
-    const int tuple_size = camp::tuple_size<decltype(argTuple)>::value;
-
-    invoke_lambda_with_args<LoopIndex>(std::forward<Data>(data),
-                                       argTuple,camp::make_idx_seq_t<tuple_size>{});
+    invoke_lambda_with_args<LoopIndex, targList>(std::forward<Data>(data));
   }
 };
